third_party/WebKit/LayoutTests/imported/wpt/html/browsers/the-window-object/Window-document-expected.txt

This is a testharness.js-based test.
FAIL Document in a browsing context assert_equals: expected "file:///common/blank.html" but got "file:///.../common/blank.html"
Harness: the test ran to completion.
